removing the main from the updater class and moving it into an execution script. Also tidying up some linter concerns and fixing some missing "self" references on logging.
